authz plugins: dynamically append authz plugins to middleware chain
Previously, authz plugins could be installed or enabled using 'docker
plugin' commands and they would show as 'enabled' in various status
reports like 'docker plugin ls' but they would not actually be
active. With this change, authz plugins are now appended to the authz
plugin chain when they are dynamically enabled so that 'docker plugin
ls' and similar do not lie.

Signed-off-by: David Sheets <dsheets@docker.com>