Close /events connections after containers shut down during daemon shutdown
Before:

If an application is watching docker /events and dockerd shutdown is requested
then the http.Server Shutdown will block until it hits the 5s timeout and will
always log the error:

```
{"component":"dockerd","error":"context canceled","level":"error","msg":"Error shutting down http server","time":"2025-11-06T11:26:52.457776878Z"}
```

For an example application which uses events and triggers this, see linuxkit's
trim-after-delete[1].

After:

We special-case the blocking /events connections and close them after we
have stopped containers (in case an orchestrator is waiting for the container
stop events).

After this patch the error is gone and the shutdown is much faster. In my test Docker
Desktop dockerd shuts down in milliseconds, rather than 5s:
```
{"component":"dockerd","level":"info","msg":"stopped gracefully after 6.178458ms","time":"2025-11-06T18:09:39.386591084Z"}
```

[1] https://github.com/linuxkit/linuxkit/blob/238449c2d6f21e163ffde04415807f9145b34557/pkg/trim-after-delete/main.go#L118

Signed-off-by: David Scott <dave@recoil.org>